requirements.txt
aniso8601==7.0.0
bcrypt==3.1.7
biopython==1.74
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Jsonpify==1.5.0
Flask-Login==0.4.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
GeoAlchemy2==0.6.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.1
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.5
